refactor: SettingCoordinating 제거 후 ProfileCoordinating으로 통합
