+++ iOS/Projects/Features/Profile/Sources/Presentation/ProfileScene/Coordinator/Protocol/ProfileCoordinating.swift
@@ -12,4 +12,10 @@ import UIKit
 public protocol ProfileCoordinating: Coordinating {
   /// 설정 창으로 이동합니다.
   func pushToSettings()
+
+  /// 로그인 창으로 이동합니다.
+  func moveToLogin()
+
+  /// 프로필 설정 화면으로 넘어갑니다.
+  func moveToProfileSettings()
 }
